Finetuning

+++ src/data_cleaning/article_processor.py
@@ -5,14 +5,15 @@ import sys
 
 from pymongo.errors import DuplicateKeyError
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+# sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
+os.chdir("../" )
 
-from data_cleaning_utils import Normalize
+from data_cleaning.data_cleaning_utils import Normalize
 
 
 class ArticleProcessor:
-  def __init__(self, scraped_collection, cleaned_collection, archive_collection, columns_to_process):
+  def __init__(self, scraped_collection=None, cleaned_collection=None, archive_collection=None, columns_to_process=[]):
     self.scraped_collection = scraped_collection
     self.cleaned_collection = cleaned_collection
     self.archive_collection = archive_collection
@@ -33,7 +34,7 @@ class ArticleProcessor:
 
   def process_article(self, article):
     # Save the _id of the article before processing it
-    article_id = article['_id']
+    article_id = article.get('_id')
 
     # Process the article
     if article is not None:
@@ -44,26 +45,30 @@ class ArticleProcessor:
         # Update the article with the cleaned text
         article[column] = clean_text
 
-      try:
-        # Try to save the processed article to the cleaned data database
-        self.cleaned_collection.insert_content(article)
-      except DuplicateKeyError:
-        print(f"Article with _id {article['_id']} already exists in the cleaned_data collection.")
+      if self.cleaned_collection is not None:
+        try:
+          # Try to save the processed article to the cleaned data database
+          self.cleaned_collection.insert_content(article)
+        except DuplicateKeyError:
+          print(f"Article with _id {article['_id']} already exists in the cleaned_data collection.")
 
-      # Move the original article to the archive database
-      self.archive_collection.insert_content(article)
+      if self.archive_collection is not None:
+        # Move the original article to the archive database
+        self.archive_collection.insert_content(article)
 
-      # Delete the original article that has been processed from the scraped data database
-      self.scraped_collection.delete_one({'_id': article_id})
+      if self.scraped_collection is not None:
+        # Delete the original article that has been processed from the scraped data database
+        self.scraped_collection.delete_one({'_id': article_id})
 
   def process_all_articles_in_collection(self):
     print("Processing all articles in the collection...")
 
-    # Fetch all articles from the current database of unclean data
-    articles = self.scraped_collection.find()
+    if self.scraped_collection is not None:
+      # Fetch all articles from the current database of unclean data
+      articles = self.scraped_collection.find()
 
-    # Process each article
-    for article in articles:
-      self.process_article(article)
+      # Process each article
+      for article in articles:
+        self.process_article(article)
 
     print("All articles have been processed.")
